[#41] style : 만들어둔 페이지 pages/friend/FriendDraftPage에  통합
